+++ lib/service/Health.dart
@@ -936,8 +936,12 @@ class Health with Service implements NotificationsListener {
   }
 
   Future<bool> clearHistory() async {
+    List<HealthHistory> history = _history;
     if (await _clearNetHistory()) {
-      await _rebuildStatus();
+      if (!ListEquality().equals(history, _history)) {
+        _notify(notifyHistoryUpdated);
+        await _rebuildStatus();
+      }
       return true;
     }
     return false;
@@ -1640,7 +1644,8 @@ class Health with Service implements NotificationsListener {
   // Vaccination
 
   bool get isVaccinated {
-    return (HealthHistory.mostRecentVaccine(_history, vaccine: HealthHistoryBlob.VaccineEffective) != null);
+    HealthHistory vaccine = HealthHistory.mostRecentVaccine(Health().history);
+    return (vaccine.blob != null) && (vaccine?.blob?.isVaccineEffective ?? false) && (vaccine.dateUtc != null) && vaccine.dateUtc.isBefore(DateTime.now().toUtc());
   }
 
   // Current Server Time