+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Deleted
+- Removed vaccine taken event handling [#715](https://github.com/rokwire/safer-illinois-app/issues/715).
 
 ## [2.11.2] - 2021-09-24
 ### Changed